Revert "Use a top level webpack config to import custom configs"
This reverts commit 980ae2efae7a2654cd098f4dd1a8be081285a04c.
